fs: implement VDI.clone
We now separate completely the "data" files from the "metadata".
The SMAPIv2 "vdi" primary key names the *metadata*, which contains
a pointer ("data") to a *data* file, either a .raw image/ standalone
.vhd or .vhd leaf.

When we snapshot, new "data" files are created with integer names
and the *metadata* is rewritten to point to the new files. No files
are renamed. Metadata names are never changed.

The scheme is based on that used by github.com/jonludlam/vhdd.