CA-272163: do not loose exception type during RPC
If the backend raises a subclass of Rpc_light_failure we were
only passing through the arguments, but not the actual exception name.
So the caller had no way to tell which exception we raised.

We need to know the correct exception when we raise the snapshot redirect
exception, otherwise XAPI just receives a generic exception it doesn't know what
to do with:
result.contents: ["Backend_error_with_backtrace",["SR_BACKEND_FAILURE",["{\"error\":\"7fe5d42c-e165-4814-8263-2408cd132a61\"

Signed-off-by: Edwin Török <edvin.torok@citrix.com>