CP-25274: SMAPIv3 interface change - Drop URI and use configuration instead
We have been using the URI parameter to store json containing the configuration.
This does not make sense, so we should stop doing that.
Instead we should use the configuration parameter, that contains a set of strings
for the same purpose. This is similar to "device-config" in the old SMAPI.

Signed-off-by: Robert Breker <robert.breker@citrix.com>

Conflicts:
	generator/lib/control.ml